# apps/api/openapi/releases/1.0.0.json
{
  "components": {
    "schemas": {
...
      "ApiErrorDetailDto": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VERSION_MISMATCH",
              "IDEMPOTENCY_KEY_REUSED",
              "RATE_LIMITED",
              "SERVICE_UNAVAILABLE",
              "INTERNAL_ERROR"
            ],
            "example": "VALIDATION_FAILED",
            "type": "string"
          },
          "message": {
            "example": "The request could not be processed.",
            "type": "string"
          },
          "requestId": {
            "example": "request_123",
            "maxLength": 128,
            "type": "string"
          }
        },
        "required": ["code", "message", "requestId"],
        "type": "object"
      },
      "ApiErrorResponseDto": {
        "additionalProperties": false,
        "properties": {
          "error": { "$ref": "#/components/schemas/ApiErrorDetailDto" }
        },
        "required": ["error"],
        "type": "object"
      },
...
      "UpdateUserRoleRequestDto": {
        "additionalProperties": false,
        "properties": {
          "role": { "enum": ["user", "admin"], "type": "string" }
        },
        "required": ["role"],
        "type": "object"
      },
      "UpdateUserStatusRequestDto": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "enum": ["pending_verification", "active", "suspended", "deleted"],
            "type": "string"
          }
        },
        "required": ["status"],
        "type": "object"
      },
      "UserListResponseDto": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/UserResponseDto" },
            "type": "array"
          },
          "page": { "example": 1, "minimum": 1, "type": "number" },
          "pageSize": {
            "example": 20,
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "total": { "example": 42, "minimum": 0, "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total"],
        "type": "object"
      },
      "UserResponseDto": {
        "additionalProperties": false,
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "email": { "format": "email", "maxLength": 254, "type": "string" },
          "emailVerifiedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": { "format": "uuid", "type": "string" },
          "role": { "enum": ["user", "admin"], "type": "string" },
          "status": {
            "enum": ["pending_verification", "active", "suspended", "deleted"],
            "type": "string"
          },
          "updatedAt": { "format": "date-time", "type": "string" },
          "version": { "minimum": 0, "type": "number" }
        },
        "required": [
          "id",
          "email",
          "status",
          "role",
          "version",
          "emailVerifiedAt",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "csrfCookie": {
        "in": "cookie",
        "name": "__Host-cs_csrf",
        "type": "apiKey"
      },
      "csrfHeader": {
        "description": "Session-bound CSRF token for unsafe cookie-auth requests.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "refreshCookie": {
        "in": "cookie",
        "name": "__Host-cs_refresh",
        "type": "apiKey"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "__Host-cs_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned Cornerstone identity and platform API contract.",
    "title": "Cornerstone API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": { "default": 1, "minimum": 1, "type": "number" }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "pending_verification",
                "active",
                "suspended",
                "deleted"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserListResponseDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "sessionCookie": [] }],
        "summary": "",
        "tags": ["Users"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "none",
          "permission": "user:list",
          "roles": ["admin"]
        }
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "operationId": "getUser",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponseDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "sessionCookie": [] }],
        "summary": "",
        "tags": ["Users"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "target",
          "permission": "user:read",
          "roles": ["admin"]
        }
      }
    },
    "/api/v1/users/{userId}/role": {
      "patch": {
        "operationId": "updateUserRole",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "description": "Strong numeric ETag, for example \"3\".",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Visible ASCII key scoped to the authenticated principal.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponseDto" }
              }
            },
            "description": "",
            "headers": {
              "ETag": { "description": "Updated strong version ETag." }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "412": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "sessionCookie": [] }
        ],
        "summary": "",
        "tags": ["Users"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "target",
          "permission": "user:update-role",
          "roles": ["admin"]
        }
      }
    },
    "/api/v1/users/{userId}/status": {
      "patch": {
        "operationId": "updateUserStatus",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "description": "Strong numeric ETag, for example \"3\".",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Visible ASCII key scoped to the authenticated principal.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponseDto" }
              }
            },
            "description": "",
            "headers": {
              "ETag": { "description": "Updated strong version ETag." }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "412": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "sessionCookie": [] }
        ],
        "summary": "",
        "tags": ["Users"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "target",
          "permission": "user:update-status",
          "roles": ["admin"]
        }
      }
    },
    "/api/v1/users/me": {
      "delete": {
        "operationId": "deleteCurrentUser",
        "parameters": [
          {
            "description": "Strong numeric ETag, for example \"3\".",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Visible ASCII key scoped to the authenticated principal.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "" },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "412": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "sessionCookie": [] }
        ],
        "summary": "",
        "tags": ["Users"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": "profile:update",
          "roles": ["user", "admin"]
        }
      }
    }
  },
  "security": [{ "sessionCookie": [] }],
  "servers": [{ "description": "API origin", "url": "/" }],
...
}